+++ .github/workflows/build-test-deploy.yaml
@@ -118,6 +118,26 @@ jobs:
       - run: ./bin/support-bundle ./examples/support-bundle/sample-supportbundle.yaml
       - run: ./bin/support-bundle https://kots.io
 
+  goreleaser-test:
+    runs-on: ubuntu-latest
+    if: startsWith(github.ref, 'refs/tags/v') != true
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+  
+      - name: Unshallow
+        run: git fetch --prune --unshallow
+  
+      - uses: actions/setup-go@v1
+        with:
+          go-version: "1.17"
+
+      - name: Run GoReleaser
+        uses: goreleaser/goreleaser-action@v2
+        with:
+          version: "v0.183.0"
+          args: build --rm-dist --snapshot --config deploy/.goreleaser.yaml
+
   goreleaser:
     runs-on: ubuntu-latest
     needs:
@@ -159,7 +179,7 @@ jobs:
       - name: Run GoReleaser
         uses: goreleaser/goreleaser-action@v2
         with:
-          version: "v0.166.1"
+          version: "v0.183.0"
           args: release --rm-dist --config deploy/.goreleaser.yaml
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
